+++ ContentPatcher/Framework/Patches/LoadPatch.cs
@@ -36,13 +36,12 @@ namespace ContentPatcher.Framework.Patches
 
         /// <summary>Update the patch data when the context changes.</summary>
         /// <param name="context">Provides access to contextual tokens.</param>
-        /// <param name="singleValueTokens">The tokens that can only contain one value.</param>
         /// <returns>Returns whether the patch data changed.</returns>
-        public override bool UpdateContext(IContext context, IDictionary<TokenName, IToken> singleValueTokens)
+        public override bool UpdateContext(IContext context)
         {
-            bool localAssetChanged = this.LocalAsset.UpdateContext(context, singleValueTokens);
+            bool localAssetChanged = this.LocalAsset.UpdateContext(context);
             this.IsValidInContext = this.ContentPack.FileExists(this.LocalAsset.Value);
-            return base.UpdateContext(context, singleValueTokens) || localAssetChanged;
+            return base.UpdateContext(context) || localAssetChanged;
         }
 
         /// <summary>Load the initial version of the asset.</summary>
@@ -56,7 +55,7 @@ namespace ContentPatcher.Framework.Patches
         }
 
         /// <summary>Get the tokens used by this patch in its fields.</summary>
-        public override IEnumerable<IToken> GetTokensUsed()
+        public override IEnumerable<TokenName> GetTokensUsed()
         {
             return base.GetTokensUsed().Union(this.LocalAsset.Tokens);
         }