+++ ContentPatcher/Framework/Tokens/VillagerHeartsToken.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using ContentPatcher.Framework.Conditions;
+using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
 using StardewValley;
 
@@ -14,7 +15,7 @@ namespace ContentPatcher.Framework.Tokens
         ** Properties
         *********/
         /// <summary>The relationships by NPC.</summary>
-        private readonly IDictionary<TokenName, string> Values = new Dictionary<TokenName, string>();
+        private readonly InvariantDictionary<string> Values = new InvariantDictionary<string>();
 
 
         /*********
@@ -22,7 +23,10 @@ namespace ContentPatcher.Framework.Tokens
         *********/
         /// <summary>Construct an instance.</summary>
         public VillagerHeartsToken()
-            : base(ConditionType.Hearts.ToString(), canHaveMultipleValues: false, requiresSubkeys: true) { }
+            : base(ConditionType.Hearts.ToString(), canHaveMultipleRootValues: false)
+        {
+            this.EnableSubkeys(required: true, canHaveMultipleValues: false);
+        }
 
         /// <summary>Update the token data when the context changes.</summary>
         /// <param name="context">The condition context.</param>
@@ -34,24 +38,25 @@ namespace ContentPatcher.Framework.Tokens
             if (this.IsValidInContext)
             {
                 foreach (KeyValuePair<string, Friendship> pair in Game1.player.friendshipData.Pairs)
-                    this.Values[new TokenName(this.Name.Key, pair.Key)] = (pair.Value.Points / NPC.friendshipPointsPerHeartLevel).ToString(CultureInfo.InvariantCulture);
+                    this.Values[pair.Key] = (pair.Value.Points / NPC.friendshipPointsPerHeartLevel).ToString(CultureInfo.InvariantCulture);
             }
         }
 
         /// <summary>Get the current subkeys (if supported).</summary>
         public override IEnumerable<TokenName> GetSubkeys()
         {
-            return this.Values.Keys;
+            foreach (string key in this.Values.Keys)
+                yield return new TokenName(this.Name.Key, key);
         }
 
         /// <summary>Get the current token values.</summary>
-        /// <param name="name">The token name to check, if applicable.</param>
-        /// <exception cref="InvalidOperationException">The key doesn't match this token, or this token require a subkeys and <paramref name="name"/> does not specify one.</exception>
-        public override IEnumerable<string> GetValues(TokenName? name = null)
+        /// <param name="name">The token name to check.</param>
+        /// <exception cref="InvalidOperationException">The key doesn't match this token, or the key does not respect <see cref="IToken.CanHaveSubkeys"/> or <see cref="IToken.RequiresSubkeys"/>.</exception>
+        public override IEnumerable<string> GetValues(TokenName name)
         {
             this.AssertTokenName(name);
 
-            if (this.Values.TryGetValue(name.Value, out string value))
+            if (this.Values.TryGetValue(name.Subkey, out string value))
                 yield return value;
         }
     }