+++ ContentPatcher/Framework/Tokens/VillagerRelationshipToken.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using ContentPatcher.Framework.Conditions;
+using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
 using StardewValley;
 
@@ -13,7 +14,7 @@ namespace ContentPatcher.Framework.Tokens
         ** Properties
         *********/
         /// <summary>The relationships by NPC.</summary>
-        private readonly IDictionary<TokenName, string> Values = new Dictionary<TokenName, string>();
+        private readonly InvariantDictionary<string> Values = new InvariantDictionary<string>();
 
 
         /*********
@@ -21,7 +22,10 @@ namespace ContentPatcher.Framework.Tokens
         *********/
         /// <summary>Construct an instance.</summary>
         public VillagerRelationshipToken()
-            : base(ConditionType.Relationship.ToString(), canHaveMultipleValues: false, requiresSubkeys: true) { }
+            : base(ConditionType.Relationship.ToString(), canHaveMultipleRootValues: false)
+        {
+            this.EnableSubkeys(required: true, canHaveMultipleValues: false);
+        }
 
         /// <summary>Update the token data when the context changes.</summary>
         /// <param name="context">The condition context.</param>
@@ -33,24 +37,25 @@ namespace ContentPatcher.Framework.Tokens
             if (this.IsValidInContext)
             {
                 foreach (KeyValuePair<string, Friendship> pair in Game1.player.friendshipData.Pairs)
-                    this.Values[new TokenName(this.Name.Key, pair.Key)] = pair.Value.Status.ToString();
+                    this.Values[pair.Key] = pair.Value.Status.ToString();
             }
         }
 
         /// <summary>Get the current subkeys (if supported).</summary>
         public override IEnumerable<TokenName> GetSubkeys()
         {
-            return this.Values.Keys;
+            foreach (string key in this.Values.Keys)
+                yield return new TokenName(this.Name.Key, key);
         }
 
         /// <summary>Get the current token values.</summary>
         /// <param name="name">The token name to check, if applicable.</param>
         /// <exception cref="InvalidOperationException">The key doesn't match this token, or this token require a subkeys and <paramref name="name"/> does not specify one.</exception>
-        public override IEnumerable<string> GetValues(TokenName? name = null)
+        public override IEnumerable<string> GetValues(TokenName name)
         {
             this.AssertTokenName(name);
 
-            if (this.Values.TryGetValue(name.Value, out string value))
+            if (this.Values.TryGetValue(name.Subkey, out string value))
                 yield return value;
         }
     }